# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Web-server for employers",
        "title": "Employers service",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/employers": {
            "post": {
                "description": "Данный роут создает сотрудника и возвращает его идентификатор в ответе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Create employer",
                "operationId": "create_employer",
                "parameters": [
                    {
                        "description": "Модель сотрудника",
                        "name": "{object}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateEmployer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сотрудник создан",
                        "schema": {
                            "$ref": "#/definitions/responses.CreateEmployer"
                        }
                    }
                }
            }
        },
        "/employers/company/{companyId}": {
            "get": {
                "description": "Данный роут позволяет получить список сотрудников по ИД компании",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Get employers by companyID",
                "operationId": "get_employer_by_company_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID компании",
                        "name": "companyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сотрудников",
                        "schema": {
                            "$ref": "#/definitions/responses.GetEmployers"
                        }
                    }
                }
            }
        },
        "/employers/department/{name}": {
            "get": {
                "description": "Данный роут позволяет получить список сотрудников по названию отдела",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Get employers by department name",
                "operationId": "get_employer_by_department_name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название отдела",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список сотрудников",
                        "schema": {
                            "$ref": "#/definitions/responses.GetEmployers"
                        }
                    }
                }
            }
        },
        "/employers/{id}": {
            "delete": {
                "description": "Данный удаляе сотрудника",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Delete employer",
                "operationId": "delete_employer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID сотрудника",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сотрудник удален",
                        "schema": {
                            "$ref": "#/definitions/responses.DeleteEmployer"
                        }
                    }
                }
            },
            "patch": {
                "description": "Данный роут позволяет изменить данные сотрудника",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Update employer by id",
                "operationId": "update_employer_by_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ид сторудника",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Модель сотрудника",
                        "name": "{object}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateEmployer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Измененная мрдель сотрудника",
                        "schema": {
                            "$ref": "#/definitions/responses.UpdateEmployer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apidto.DepartmentDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "apidto.EmployerDTO": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department": {
                    "$ref": "#/definitions/apidto.DepartmentDTO"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "passport": {
                    "$ref": "#/definitions/apidto.PassportDTO"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "apidto.PassportDTO": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "requests.CreateEmployer": {
            "type": "object",
            "required": [
                "company_id",
                "department",
                "name",
                "passport",
                "phone",
                "surname"
            ],
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department": {
                    "$ref": "#/definitions/apidto.DepartmentDTO"
                },
                "name": {
                    "type": "string"
                },
                "passport": {
                    "$ref": "#/definitions/apidto.PassportDTO"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "requests.UpdateEmployer": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department_name": {
                    "type": "string"
                },
                "department_phone": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "passport_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "responses.CreateEmployer": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "responses.DeleteEmployer": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.GetEmployers": {
            "type": "object",
            "properties": {
                "employers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apidto.EmployerDTO"
                    }
                }
            }
        },
        "responses.UpdateEmployer": {
            "type": "object",
            "properties": {
                "employer": {
                    "$ref": "#/definitions/apidto.EmployerDTO"
                }
            }
        }
    }
}